api/requirements.txt
fastapi
uvicorn
pydantic
pyyaml
redis

